Extract codec.ts
