+++ packages/iov-bcp-types/src/connection.ts
@@ -5,7 +5,8 @@ import { Stream } from "xstream";
 import { ChainId, PostableBytes, PublicKeyBundle } from "@iov/base-types";
 import { ValueAndUpdates } from "@iov/stream";
 
-import { Address, SignedTransaction, TransactionId, TxCodec } from "./signables";
+import { TxCodec } from "./codec";
+import { Address, SignedTransaction, TransactionId } from "./signables";
 import { Amount, Nonce, TokenTicker, UnsignedTransaction } from "./transactions";
 
 /*
