+++ packages/iov-bcp-types/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./atomicswap";
+export * from "./codec";
 export * from "./connection";
 export * from "./signables";
 export * from "./transactions";
