+++ packages/iov-bcp-types/types/index.d.ts
@@ -1,4 +1,5 @@
 export * from "./atomicswap";
+export * from "./codec";
 export * from "./connection";
 export * from "./signables";
 export * from "./transactions";
