+++ packages/iov-bcp-types/types/signables.d.ts
@@ -1,5 +1,5 @@
 import { As } from "type-tagger";
-import { ChainId, PostableBytes, PublicKeyBundle, SignatureBytes } from "@iov/base-types";
+import { PublicKeyBundle, SignatureBytes } from "@iov/base-types";
 import { Nonce, UnsignedTransaction } from "./transactions";
 /**
  * A printable transaction ID in a blockchain-specific format.
@@ -35,21 +35,3 @@ export interface SignedTransaction<T extends UnsignedTransaction = UnsignedTrans
 }
 /** A codec specific address encoded as a string */
 export declare type Address = string & As<"address">;
-export interface TxReadCodec {
-    /** parseBytes will recover bytes from the blockchain into a format we can use */
-    readonly parseBytes: (bytes: PostableBytes, chainId: ChainId) => SignedTransaction;
-    /** chain-dependent way to calculate address from key */
-    readonly keyToAddress: (key: PublicKeyBundle) => Address;
-    /** chain-dependent validation of address */
-    readonly isValidAddress: (address: string) => boolean;
-}
-/** TxCodec knows how to convert Transactions to bytes for a given blockchain */
-export interface TxCodec extends TxReadCodec {
-    /** these are the bytes we create to add a signature */
-    /** they often include nonce and chainID, but not other signatures */
-    readonly bytesToSign: (tx: UnsignedTransaction, nonce: Nonce) => SigningJob;
-    /** bytesToPost includes the raw transaction appended with the various signatures */
-    readonly bytesToPost: (tx: SignedTransaction) => PostableBytes;
-    /** identifier is usually some sort of hash of bytesToPost, chain-dependent */
-    readonly identifier: (tx: SignedTransaction) => TransactionId;
-}
